Reorganize codebase into domain packages
Extracts flat root-level modules into four domain packages with
well-defined contracts, laying the groundwork for a microservice-style
architecture.

- ai/: ChatGPTHandler + DENJAMIN_ROLE config (OpenAI concern)
- db/: PointsRepository + connection helper (persistence concern)
- discord_utils/: message/image formatting utilities (Discord-to-AI transforms)
- bot/: MyBot client, on_message handler, slash commands (Discord concern)

denjamin.py becomes a slim composition root (~12 lines) that wires
dependencies together and runs the bot.

Closes #7
