+++ bot/__init__.py
@@ -1,0 +1,1 @@
+from bot.client import create_bot
