+++ bot/commands/__init__.py
@@ -1,0 +1,1 @@
+from bot.commands.points import register_points_commands
