+++ db/__init__.py
@@ -1,0 +1,1 @@
+from db.points import PointsRepository
