+++ db/connection.py
@@ -1,0 +1,5 @@
+import psycopg2
+
+
+def get_db_conn(database_url):
+    return psycopg2.connect(database_url)
